style(pubspec.yaml): enforce consistent formatting
This update is concentrated on refining formatting in the 'pubspec.yaml' file for better readability and consistent code style. Main points of this commit include: standardizing quotes, removing excessive blank lines for brevity, and adhering to uniform spaces with types in 'spider' configuration. These updates do not impact functionalities, they simply make the code cleaner and easier to read.